requirements.txt
customtkinter>=5.2
Pillow>=10.0
send2trash>=1.8
watchdog>=4.0
